Code refactor: - Move functions from tasks.py to utils.py - Create separate function for updatinf the activity status and message - Fix lint check errors
